Initial commit
This commit represents a wholesale rewrite of the `install-wasi-sdk`
action [here]. After discussion with the author of that crate, we agreed
that license requirements within the Bytecode Alliance and the need to
run the action on GitHub runners of any OS made a rewrite the best
option. Though this repository is now a completely new thing, I wanted
to recognize previous work as the inspiration for this. Thank you!

[here]: https://github.com/konsumer/install-wasi-sdk